added further tuning schemes

+++ dreisam_moehlin_neumagen/steady-state/drainage_schoenberg-no-flow_black-forest-no-flow/generate_boundary_conditions.py
@@ -177,7 +177,7 @@ def main(offset, plot):
         v.attrs.update(long_name="location of constant head boundary condition of the porous aquifer", units="-")
 
         v = f.create_variable(
-            "constant_head_porous_aquifer", ("x", "y"), np.int32, compression="gzip", compression_opts=1
+            "constant_head_porous_aquifer", ("x", "y"), np.float32, compression="gzip", compression_opts=1
         )
         v[:, :] = constant_head_porous_aquifer[:, :] - offset
         v.attrs.update(long_name="constant head boundary condition of the porous aquifer", units="m a.s.l.")
@@ -189,7 +189,7 @@ def main(offset, plot):
         v.attrs.update(long_name="location of schoenberg boundary condition", units="-")
 
         v = f.create_variable(
-            "constant_head_schoenberg", ("x", "y"), np.int32, compression="gzip", compression_opts=1
+            "constant_head_schoenberg", ("x", "y"), np.float32, compression="gzip", compression_opts=1
         )
         v[:, :] = constant_head_schoenberg[:, :] - offset
         v.attrs.update(long_name="schoenberg boundary condition", units="m a.s.l.")
@@ -201,7 +201,7 @@ def main(offset, plot):
         v.attrs.update(long_name="location of the boundary condition of the fissured aquifer", units="-")
 
         v = f.create_variable(
-            "constant_head_black_forest", ("x", "y"), np.int32, compression="gzip", compression_opts=1
+            "constant_head_black_forest", ("x", "y"), np.float32, compression="gzip", compression_opts=1
         )
         v[:, :] = constant_head_black_forest[:, :] - offset
         v.attrs.update(long_name="boundary condition of the fissured auifer", units="m a.s.l.")
